requirements.txt
pandas>=1.3.0
requests>=2.26.0
python-dotenv>=0.19.0
mysql-connector-python>=8.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
python-dateutil>=2.8.0